+++ @yundeving/story-book-demo-ui/ui/components/Button/Button.type.ts
@@ -1,10 +1,11 @@
 import { type SizeType, type VariantType } from "../common.type";
-
-export type ButtonStatus = "default" | "loading" | "success" | "error";
+import { type ReactNode } from "react";
 
 export interface ButtonProps extends React.ComponentProps<"button"> {
   label?: string;
   variant?: VariantType | "ghost";
   size?: SizeType;
-  status?: ButtonStatus;
+  isLoading?: boolean;
+  leftIcon?: ReactNode;
+  rightIcon?: ReactNode;
 }