+++ @yundeving/story-book-demo-ui/ui/components/IconButton/IconButton.type.ts
@@ -1,6 +1,0 @@
-import { type ButtonProps } from "../Button/Button.type";
-
-export interface IconButtonProps extends ButtonProps {
-  icon: React.ReactNode;
-  onlyIcon?: boolean;
-}